1) Added AbstractAmazon::getRequestDate & AbstractAmazon::setRequestDate for testing services with signature 2) created a draft test for Service\Amazon\S3\S3::createBucket
(deleted S3TestTemp.php)
